Tune dice token matching and postfix operators processing
Dice token must not contain any spaces (4d6h is valid; 4d6 h, 4 d6h, 4 d6 h are not. The latter is a valid composition operator notation)
Postfix operators (the ones with 0 right arity) are treated like operands, since their invocation is immediate. This means that they now can be chained with following infix operators. As a result, composition operators can be used without brackets (6 d4 s + 1 is valid and is equal to 6d4 + 1). They still has to be surrounded with brackets to be used with preceding prefix operators (- 3 d4 s equals to -3d4 which is not valid; -(3 d4 l) is a valid form)