Make operators be aware of their position within the expression
In regular shunting yard it is expected that there are never more operands than the operators expect to have. For example, "1-1 2" would yield -1, because the operator blindly accepts the topmost 2 operands from the stack (1 and 2) instead of filtering the excessive operand out.
Now operands do keep their position info and are provided with adjacent operands
Previously only limited to delayed operators, which are now removed